[Tech Debt] Reorganize JS files
