+++ js/components/device_demographics/BrowsersChart.js
@@ -2,35 +2,35 @@ import React, { useRef, useEffect } from "react";
 import PropTypes from "prop-types";
 import d3 from "d3";
 
-import renderBlock from "../../js/lib/renderblock";
+import renderBlock from "../../lib/chart_helpers/renderblock";
 
-function TopSourceMedia({ dataHrefBase }) {
-  const dataURL = `${dataHrefBase}/top-session-source-medium-30-days.json`;
+function BrowsersChart({ dataHrefBase }) {
+  const dataURL = `${dataHrefBase}/browsers.json`;
   const ref = useRef(null);
 
   useEffect(() => {
-    const initTopSourceMediaChart = async () => {
+    const initBrowsersChart = async () => {
       const result = await d3
         .select(ref.current)
         .datum({
           source: dataURL,
           block: ref.current,
         })
-        .call(renderBlock.buildCompactBarChart("session_source_medium"));
+        .call(renderBlock.buildCompactBarChart("browser"));
       return result;
     };
-    initTopSourceMediaChart().catch(console.error);
+    initBrowsersChart().catch(console.error);
   }, []);
 
   return (
-    <figure id="chart_session_source_medium" ref={ref}>
+    <figure id="chart_browsers" ref={ref}>
       <div className="data bar-chart"></div>
     </figure>
   );
 }
 
-TopSourceMedia.propTypes = {
+BrowsersChart.propTypes = {
   dataHrefBase: PropTypes.string.isRequired,
 };
 
-export default TopSourceMedia;
+export default BrowsersChart;
