+++ js/components/device_demographics/DeviceDemographics.js
