+++ js/components/device_demographics/index.js
