+++ js/components/top_pages/TopPagesRealtime.js
@@ -2,17 +2,17 @@ import React, { useRef, useEffect } from "react";
 import PropTypes from "prop-types";
 import d3 from "d3";
 
-import renderBlock from "../../js/lib/renderblock";
-import { exceptions, titleExceptions } from "../../js/lib/exceptions";
-import barChart from "../../js/lib/barchart";
-import formatters from "../../js/lib/formatters";
+import renderBlock from "../../lib/chart_helpers/renderblock";
+import { titleExceptions } from "../../lib/exceptions";
+import barChart from "../../lib/chart_helpers/barchart";
+import formatters from "../../lib/chart_helpers/formatters";
 
-function TopPagesHistorical({ dataHrefBase, reportFileName }) {
+function TopPagesRealtime({ dataHrefBase, reportFileName }) {
   const dataURL = `${dataHrefBase}/${reportFileName}`;
   const ref = useRef(null);
 
   useEffect(() => {
-    const initHistoricalChart = async () => {
+    const initRealtimeChart = async () => {
       const result = await d3
         .select(ref.current)
         .datum({
@@ -24,26 +24,18 @@ function TopPagesHistorical({ dataHrefBase, reportFileName }) {
             .loadAndRender()
             .transform((d) => d.data)
             .on("render", (selection) => {
-              // turn the labels into links
               selection
                 .selectAll(".label")
                 .each(function (d) {
                   d.text = this.innerText;
                 })
                 .html("")
-                .append("a")
-                .attr("target", "_blank")
-                .attr("rel", "noopener")
-                .attr(
-                  "href",
-                  (d) => exceptions[d.domain] || `http://${d.domain}`,
-                )
-                .text((d) => titleExceptions[d.domain] || d.domain);
+                .text((d) => titleExceptions[d.page] || d.page_title);
             })
             .render(
               barChart()
-                .label((d) => d.domain)
-                .value((d) => +d.visits)
+                .label((d) => d.page_title)
+                .value((d) => +d.active_visitors)
                 .scale((values) =>
                   d3.scale
                     .linear()
@@ -55,19 +47,24 @@ function TopPagesHistorical({ dataHrefBase, reportFileName }) {
         );
       return result;
     };
-    initHistoricalChart().catch(console.error);
-  });
+    initRealtimeChart().catch(console.error);
+  }, []);
 
   return (
-    <figure className="bar-chart__top-pages" data-source={dataURL} ref={ref}>
+    <figure
+      className="top-pages__realtime"
+      data-source={dataURL}
+      data-refresh="15"
+      ref={ref}
+    >
       <div className="data bar-chart"></div>
     </figure>
   );
 }
 
-TopPagesHistorical.propTypes = {
+TopPagesRealtime.propTypes = {
   dataHrefBase: PropTypes.string.isRequired,
   reportFileName: PropTypes.string.isRequired,
 };
 
-export default TopPagesHistorical;
+export default TopPagesRealtime;
