+++ js/lib/chart_helpers/barchart.js
