+++ js/lib/chart_helpers/formatters.js
