+++ js/lib/chart_helpers/renderblock.js
