+++ js/lib/chart_helpers/timeseries.js
