+++ js/lib/chart_helpers/transformers.js
