+++ js/lib/react_setup.js
@@ -1,11 +1,11 @@
-require("../../components/agency_select");
-require("../../components/data_downloads");
-require("../../components/device_demographics");
-require("../../components/engagement");
-require("../../components/locations_and_languages");
-require("../../components/realtime_visitors");
-require("../../components/sessions_30_days");
-require("../../components/top_pages");
-require("../../components/top_downloads");
-require("../../components/traffic_sources");
-require("../../components/visitors_30_days");
+require("../components/agency_select");
+require("../components/data_downloads");
+require("../components/device_demographics");
+require("../components/engagement");
+require("../components/locations_and_languages");
+require("../components/realtime_visitors");
+require("../components/sessions_30_days");
+require("../components/top_pages");
+require("../components/top_downloads");
+require("../components/traffic_sources");
+require("../components/visitors_30_days");
