+++ sass/_uswds-theme-custom-styles.scss
@@ -108,7 +108,7 @@ figure figure {
   }
 }
 
-*[data-block] {
+.data {
   opacity: 1;
   // transition opacity over a fraction of a second, with a delay
   // (the delay guards against flashes when content switches really quickly
